Implemented cancel action
When a user cancels the Sign2Pay payment screen, the order is now
marked as Cancelled and a new cart is created based on the contents
of the old order. This allows the user to choose a different
payment method or try again.

Also, default status is now Payment Pending.